ml_dynamic_app/requirements.txt
Flask>=2.0.0
joblib>=1.1.0
requests>=2.25.0
transformers>=4.20.0
torch>=1.12.0
python-dotenv>=0.19.0
flask-cors>=3.0.10
